requirements.txt
torch~=2.0.1
numpy~=1.26.2
scipy~=1.10.1
matplotlib~=3.6.3
scikit-learn~=1.2.1
torchvision~=0.15.2
Pillow~=9.3.0